fix freeze money interactor, add subscribe to offer method to offer s…

+++ bambook_api/app/interactors/freeze_money.rb
@@ -12,7 +12,8 @@ class FreezeMoney
       context.user.wallet.update(available_money: rest_money, frozen_money: frozen_money)
       context.operation_code = 2
     else
-      context.offer.destroy
+      # binding.pry
+      context.offer.destroy if context.offer.user_id == context.user.id
       context.errors = 'dont have enough money'
       context.fail!
     end
